chore: move mapping to usecase

+++ src/Applications/use_case/_test/DeleteReplyUseCase.test.js
@@ -4,11 +4,13 @@ const DeleteReplyUseCase = require('../DeleteReplyUseCase');
 describe('DeleteReplyUseCase', () => {
   it('should delete sucessfully', async () => {
     const replyRepo = new ReplyRepository();
+    replyRepo.validateOwnership = jest.fn(() => Promise.resolve());
     replyRepo.deleteReply = jest.fn(() => Promise.resolve());
 
     const usecase = new DeleteReplyUseCase(replyRepo);
     await usecase.execute('1', '2', '3');
 
-    expect(replyRepo.deleteReply).toBeCalledWith('1', '2', '3');
+    expect(replyRepo.validateOwnership).toBeCalledWith('1', '2', '3');
+    expect(replyRepo.deleteReply).toBeCalledWith('2');
   });
 });
